// frontend/src/api/generated/models/UpdateOperationalExpenseDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExpenseCategory } from './ExpenseCategory';
import {
    ExpenseCategoryFromJSON,
    ExpenseCategoryFromJSONTyped,
    ExpenseCategoryToJSON,
} from './ExpenseCategory';

/**
 * 
 * @export
 * @interface UpdateOperationalExpenseDTO
 */
export interface UpdateOperationalExpenseDTO {
    /**
     * 
     * @type {number}
     * @memberof UpdateOperationalExpenseDTO
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateOperationalExpenseDTO
     */
    insertDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateOperationalExpenseDTO
     */
    costDate?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateOperationalExpenseDTO
     */
    amount?: number | null;
    /**
     * 
     * @type {ExpenseCategory}
     * @memberof UpdateOperationalExpenseDTO
     */
    category?: ExpenseCategory | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateOperationalExpenseDTO
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateOperationalExpenseDTO
     */
    invoiceNumber?: string | null;
}



/**
 * Check if a given object implements the UpdateOperationalExpenseDTO interface.
 */
export function instanceOfUpdateOperationalExpenseDTO(value: object): value is UpdateOperationalExpenseDTO {
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function UpdateOperationalExpenseDTOFromJSON(json: any): UpdateOperationalExpenseDTO {
    return UpdateOperationalExpenseDTOFromJSONTyped(json, false);
}

export function UpdateOperationalExpenseDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateOperationalExpenseDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'insertDate': json['insertDate'] == null ? undefined : json['insertDate'],
        'costDate': json['costDate'] == null ? undefined : json['costDate'],
        'amount': json['amount'] == null ? undefined : json['amount'],
        'category': json['category'] == null ? undefined : ExpenseCategoryFromJSON(json['category']),
        'description': json['description'] == null ? undefined : json['description'],
        'invoiceNumber': json['invoiceNumber'] == null ? undefined : json['invoiceNumber'],
    };
}

export function UpdateOperationalExpenseDTOToJSON(value?: UpdateOperationalExpenseDTO | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'insertDate': value['insertDate'],
        'costDate': value['costDate'],
        'amount': value['amount'],
        'category': ExpenseCategoryToJSON(value['category']),
        'description': value['description'],
        'invoiceNumber': value['invoiceNumber'],
    };
}